refactor(agent): 提取运行时补丁管理和循环钩子构建逻辑
将 LoopRunner 中的运行时补丁管理逻辑提取到 RuntimePatchManager 类，负责：
- 管理运行时状态（skill、tool policy、context overlay、discovered tools、hooks、patch history）
- 派生和应用 RuntimePatch
- 构建运行时上下文快照
- 管理 Skill 激活/清除
- 管理工具发现和目录消息同步
- 清理 turn-scoped 状态

将循环钩子构建逻辑提取到 buildLoopConfig 函数，负责：
- 构建 AgentLoopConfig 对象（包含所有 hooks）
- 统一 JSONL 持久化模式

同时清理了过时的上下文模块文档，并修复了类型导入和测试文件中的类型声明。
