+++ src/agent/state/TurnState.ts
@@ -4,7 +4,7 @@ import type { ToolCatalog } from '../../tools/catalog/index.js';
 import type { ToolRegistry } from '../../tools/registry/ToolRegistry.js';
 import type { ConfirmationHandler } from '../../tools/types/ExecutionTypes.js';
 import type { PermissionMode } from '../../types/common.js';
-import type { BackgroundAgentManager } from '../subagents/BackgroundAgentManager.js';
+import type { IBackgroundAgentManager } from '../types.js';
 
 export type LlmToolDefinition = {
   name: string;
@@ -33,7 +33,7 @@ export interface LoopExecutionContext {
   contextSnapshot?: ContextSnapshot;
   skillActivationPaths?: string[];
   confirmationHandler?: ConfirmationHandler;
-  backgroundAgentManager?: BackgroundAgentManager;
+  backgroundAgentManager?: IBackgroundAgentManager;
   toolRegistry?: ToolRegistry;
   toolCatalog?: ToolCatalog;
   discoveredTools?: string[];
